Update artists.css
I'm new to Github, have no idea how this works, but I added a bunch of verified artists accounts that we never added due to the current CSS lagging out a lot. Also, is it possible to use the CSS to apply gold names to any approved submitters instead of doing this, since we've been adding any verified artist accounts to the approved submitters anyway.